tests/wpt/web-platform-tests/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/img-replaced-box-while-loading.html

<!doctype html>
<title>Images don't render as a non-replaced inline while loading, even when there's no concrete size specified</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1472637">
<style>
  img {
    min-width: 1000px;
  }
</style>
<img alt="T">
<script>
// Do an async test off the onload handler to avoid waiting for the load even for too long unnecessarily.
let t = async_test("Loading images should get a replaced box, even without specified size");
onload = t.step_func(function() {
  const image = document.querySelector("img");
  // Use the trickle pipe to have 100 seconds until the image actually loads,
  // that should be enough to run the test.
  image.src = "../../../../../images/blue.png?pipe=trickle(d100)";
  t.step_timeout(t.step_func_done(function() {
    assert_equals(
      image.offsetWidth,
      1000,
    );
  }), 0);
});
</script>
